Move FunSpecs relating to csv exporting to their own file.
